Thunks/gen: Move definition of GenerateThunkLibsAction into gen.cpp
